backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
openai==1.54.0
python-dotenv==1.0.0
faiss-cpu==1.7.4
pandas==2.1.4
numpy==1.26.2
tiktoken==0.5.2
requests==2.31.0
